+++ types/document.d.ts
@@ -24,9 +24,6 @@ declare module 'mongoose' {
     /** This documents _id. */
     _id: T;
 
-    /** This documents __v. */
-    __v?: any;
-
     /** Assert that a given path or paths is populated. Throws an error if not populated. */
     $assertPopulated<Paths = {}>(path: string | string[], values?: Partial<Paths>): Omit<this, keyof Paths> & Paths;
 
@@ -138,7 +135,7 @@ declare module 'mongoose' {
      * Takes a populated field and returns it to its unpopulated state. If called with
      * no arguments, then all populated fields are returned to their unpopulated state.
      */
-    depopulate(path?: string | string[]): this;
+    depopulate<Paths = {}>(path?: string | string[]): MergeType<this, Paths>;
 
     /**
      * Returns the list of paths that have been directly modified. A direct
@@ -259,11 +256,14 @@ declare module 'mongoose' {
     set(value: string | Record<string, any>): this;
 
     /** The return value of this method is used in calls to JSON.stringify(doc). */
+    toJSON(options?: ToObjectOptions & { flattenMaps?: true }): FlattenMaps<Require_id<DocType>>;
+    toJSON(options: ToObjectOptions & { flattenMaps: false }): Require_id<DocType>;
     toJSON<T = Require_id<DocType>>(options?: ToObjectOptions & { flattenMaps?: true }): FlattenMaps<T>;
     toJSON<T = Require_id<DocType>>(options: ToObjectOptions & { flattenMaps: false }): T;
 
     /** Converts this document into a plain-old JavaScript object ([POJO](https://masteringjs.io/tutorials/fundamentals/pojo)). */
-    toObject<T = Require_id<DocType>>(options?: ToObjectOptions): Require_id<T>;
+    toObject(options?: ToObjectOptions): Require_id<DocType>;
+    toObject<T>(options?: ToObjectOptions): Require_id<T>;
 
     /** Clears the modified state on the specified path. */
     unmarkModified<T extends keyof DocType>(path: T): void;