+++ types/inferrawdoctype.d.ts
@@ -1,4 +1,5 @@
 import {
+  IsPathRequired,
   IsSchemaTypeFromBuiltinClass,
   RequiredPaths,
   OptionalPaths,
@@ -14,7 +15,9 @@ declare module 'mongoose' {
     [
     K in keyof (RequiredPaths<DocDefinition, TSchemaOptions['typeKey']> &
     OptionalPaths<DocDefinition, TSchemaOptions['typeKey']>)
-    ]: ObtainRawDocumentPathType<DocDefinition[K], TSchemaOptions['typeKey']>;
+    ]: IsPathRequired<DocDefinition[K], TSchemaOptions['typeKey']> extends true
+      ? ObtainRawDocumentPathType<DocDefinition[K], TSchemaOptions['typeKey']>
+      : ObtainRawDocumentPathType<DocDefinition[K], TSchemaOptions['typeKey']> | null;
   }, TSchemaOptions>;
 
   /**
@@ -88,8 +91,8 @@ declare module 'mongoose' {
             IfEquals<PathValueType, String> extends true ? PathEnumOrString<Options['enum']> :
               PathValueType extends NumberSchemaDefinition ? Options['enum'] extends ReadonlyArray<any> ? Options['enum'][number] : number :
                 IfEquals<PathValueType, Schema.Types.Number> extends true ? number :
-                  PathValueType extends DateSchemaDefinition ? Date :
-                    IfEquals<PathValueType, Schema.Types.Date> extends true ? Date :
+                  PathValueType extends DateSchemaDefinition ? NativeDate :
+                    IfEquals<PathValueType, Schema.Types.Date> extends true ? NativeDate :
                       PathValueType extends typeof Buffer | 'buffer' | 'Buffer' | typeof Schema.Types.Buffer ? Buffer :
                         PathValueType extends BooleanSchemaDefinition ? boolean :
                           IfEquals<PathValueType, Schema.Types.Boolean> extends true ? boolean :